+++ cli/src/lib.rs
@@ -101,6 +101,9 @@ pub enum Command {
         /// verifiable builds. Only works for debian-based images.
         #[clap(value_enum, short, long, default_value = "none")]
         bootstrap: BootstrapMode,
+        /// Environment variables to pass into the docker container
+        #[clap(short, long, required = false)]
+        env: Vec<String>,
         /// Arguments to pass to the underlying `cargo build-bpf` command
         #[clap(required = false, last = true)]
         cargo_args: Vec<String>,
@@ -141,6 +144,9 @@ pub enum Command {
         /// verifiable builds. Only works for debian-based images.
         #[clap(value_enum, short, long, default_value = "none")]
         bootstrap: BootstrapMode,
+        /// Environment variables to pass into the docker container
+        #[clap(short, long, required = false)]
+        env: Vec<String>,
         /// Arguments to pass to the underlying `cargo build-bpf` command.
         #[clap(required = false, last = true)]
         cargo_args: Vec<String>,
@@ -172,6 +178,9 @@ pub enum Command {
         #[clap(long)]
         run: Vec<String>,
         args: Vec<String>,
+        /// Environment variables to pass into the docker container
+        #[clap(short, long, required = false)]
+        env: Vec<String>,
         /// Arguments to pass to the underlying `cargo build-bpf` command.
         #[clap(required = false, last = true)]
         cargo_args: Vec<String>,
@@ -237,6 +246,9 @@ pub enum Command {
     Publish {
         /// The name of the program to publish.
         program: String,
+        /// Environment variables to pass into the docker container
+        #[clap(short, long, required = false)]
+        env: Vec<String>,
         /// Arguments to pass to the underlying `cargo build-bpf` command.
         #[clap(required = false, last = true)]
         cargo_args: Vec<String>,
@@ -264,6 +276,9 @@ pub enum Command {
         /// no "CHECK" comments where normally required
         #[clap(long)]
         skip_lint: bool,
+        /// Environment variables to pass into the docker container
+        #[clap(short, long, required = false)]
+        env: Vec<String>,
         /// Arguments to pass to the underlying `cargo build-bpf` command.
         #[clap(required = false, last = true)]
         cargo_args: Vec<String>,
@@ -292,6 +307,9 @@ pub enum IdlCommand {
         program_id: Pubkey,
         #[clap(short, long)]
         filepath: String,
+    },
+    Close {
+        program_id: Pubkey,
     },
     /// Writes an IDL into a buffer account. This can be used with SetBuffer
     /// to perform an upgrade.
@@ -387,6 +405,7 @@ pub fn entry(opts: Opts) -> Result<()> {
             docker_image,
             bootstrap,
             cargo_args,
+            env,
             skip_lint,
             no_docs,
         } => build(
@@ -401,6 +420,7 @@ pub fn entry(opts: Opts) -> Result<()> {
             bootstrap,
             None,
             None,
+            env,
             cargo_args,
             no_docs,
         ),
@@ -410,6 +430,7 @@ pub fn entry(opts: Opts) -> Result<()> {
             solana_version,
             docker_image,
             bootstrap,
+            env,
             cargo_args,
         } => verify(
             &opts.cfg_override,
@@ -418,6 +439,7 @@ pub fn entry(opts: Opts) -> Result<()> {
             solana_version,
             docker_image,
             bootstrap,
+            env,
             cargo_args,
         ),
         Command::Clean => clean(&opts.cfg_override),
@@ -442,6 +464,7 @@ pub fn entry(opts: Opts) -> Result<()> {
             detach,
             run,
             args,
+            env,
             cargo_args,
             skip_lint,
         } => test(
@@ -453,6 +476,7 @@ pub fn entry(opts: Opts) -> Result<()> {
             detach,
             run,
             args,
+            env,
             cargo_args,
         ),
         #[cfg(feature = "dev")]
@@ -466,20 +490,23 @@ pub fn entry(opts: Opts) -> Result<()> {
         Command::Login { token } => login(&opts.cfg_override, token),
         Command::Publish {
             program,
+            env,
             cargo_args,
             skip_build,
-        } => publish(&opts.cfg_override, program, cargo_args, skip_build),
+        } => publish(&opts.cfg_override, program, env, cargo_args, skip_build),
         Command::Keys { subcmd } => keys(&opts.cfg_override, subcmd),
         Command::Localnet {
             skip_build,
             skip_deploy,
             skip_lint,
+            env,
             cargo_args,
         } => localnet(
             &opts.cfg_override,
             skip_build,
             skip_deploy,
             skip_lint,
+            env,
             cargo_args,
         ),
         Command::Account {
@@ -790,6 +817,7 @@ pub fn build(
     bootstrap: BootstrapMode,
     stdout: Option<File>, // Used for the package registry server.
     stderr: Option<File>, // Used for the package registry server.
+    env_vars: Vec<String>,
     cargo_args: Vec<String>,
     no_docs: bool,
 ) -> Result<()> {
@@ -835,6 +863,7 @@ pub fn build(
             &build_config,
             stdout,
             stderr,
+            env_vars,
             cargo_args,
             skip_lint,
             no_docs,
@@ -848,6 +877,7 @@ pub fn build(
             &build_config,
             stdout,
             stderr,
+            env_vars,
             cargo_args,
             skip_lint,
             no_docs,
@@ -861,6 +891,7 @@ pub fn build(
             &build_config,
             stdout,
             stderr,
+            env_vars,
             cargo_args,
             skip_lint,
             no_docs,
@@ -881,6 +912,7 @@ fn build_all(
     build_config: &BuildConfig,
     stdout: Option<File>, // Used for the package registry server.
     stderr: Option<File>, // Used for the package registry server.
+    env_vars: Vec<String>,
     cargo_args: Vec<String>,
     skip_lint: bool,
     no_docs: bool,
@@ -898,6 +930,7 @@ fn build_all(
                     build_config,
                     stdout.as_ref().map(|f| f.try_clone()).transpose()?,
                     stderr.as_ref().map(|f| f.try_clone()).transpose()?,
+                    env_vars.clone(),
                     cargo_args.clone(),
                     skip_lint,
                     no_docs,
@@ -920,6 +953,7 @@ fn build_cwd(
     build_config: &BuildConfig,
     stdout: Option<File>,
     stderr: Option<File>,
+    env_vars: Vec<String>,
     cargo_args: Vec<String>,
     skip_lint: bool,
     no_docs: bool,
@@ -937,6 +971,7 @@ fn build_cwd(
             stdout,
             stderr,
             skip_lint,
+            env_vars,
             cargo_args,
             no_docs,
         ),
@@ -953,6 +988,7 @@ fn build_cwd_verifiable(
     stdout: Option<File>,
     stderr: Option<File>,
     skip_lint: bool,
+    env_vars: Vec<String>,
     cargo_args: Vec<String>,
     no_docs: bool,
 ) -> Result<()> {
@@ -975,6 +1011,7 @@ fn build_cwd_verifiable(
         build_config,
         stdout,
         stderr,
+        env_vars,
         cargo_args,
     );
 
@@ -1014,6 +1051,7 @@ fn build_cwd_verifiable(
     result
 }
 
+#[allow(clippy::too_many_arguments)]
 fn docker_build(
     cfg: &WithPath<Config>,
     container_name: &str,
@@ -1021,6 +1059,7 @@ fn docker_build(
     build_config: &BuildConfig,
     stdout: Option<File>,
     stderr: Option<File>,
+    env_vars: Vec<String>,
     cargo_args: Vec<String>,
 ) -> Result<()> {
     let binary_name = Manifest::from_path(&cargo_toml)?.lib_name()?;
@@ -1074,6 +1113,7 @@ fn docker_build(
             binary_name,
             stdout,
             stderr,
+            env_vars,
             cargo_args,
         )
     });
@@ -1137,6 +1177,7 @@ fn docker_build_bpf(
     binary_name: String,
     stdout: Option<File>,
     stderr: Option<File>,
+    env_vars: Vec<String>,
     cargo_args: Vec<String>,
 ) -> Result<()> {
     let manifest_path =
@@ -1154,6 +1195,13 @@ fn docker_build_bpf(
             "exec",
             "--env",
             "PATH=/root/.local/share/solana/install/active_release/bin:/root/.cargo/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin",
+        ])
+        .args(env_vars
+            .iter()
+            .map(|x| ["--env", x.as_str()])
+            .collect::<Vec<[&str; 2]>>()
+            .concat())
+        .args([
             container_name,
             "cargo",
             "build-bpf",
@@ -1292,6 +1340,7 @@ fn _build_cwd(
     Ok(())
 }
 
+#[allow(clippy::too_many_arguments)]
 fn verify(
     cfg_override: &ConfigOverride,
     program_id: Pubkey,
@@ -1299,6 +1348,7 @@ fn verify(
     solana_version: Option<String>,
     docker_image: Option<String>,
     bootstrap: BootstrapMode,
+    env_vars: Vec<String>,
     cargo_args: Vec<String>,
 ) -> Result<()> {
     // Change to the workspace member directory, if needed.
@@ -1324,6 +1374,7 @@ fn verify(
         bootstrap,                                             // bootstrap docker image
         None,                                                  // stdout
         None,                                                  // stderr
+        env_vars,
         cargo_args,
         false,
     )?;
@@ -1517,7 +1568,9 @@ fn fetch_idl(cfg_override: &ConfigOverride, idl_addr: Pubkey) -> Result<Idl> {
     let mut d: &[u8] = &account.data[8..];
     let idl_account: IdlAccount = AnchorDeserialize::deserialize(&mut d)?;
 
-    let mut z = ZlibDecoder::new(&idl_account.data[..]);
+    let compressed_len: usize = idl_account.data_len.try_into().unwrap();
+    let compressed_bytes = &account.data[44..44 + compressed_len];
+    let mut z = ZlibDecoder::new(compressed_bytes);
     let mut s = Vec::new();
     z.read_to_end(&mut s)?;
     serde_json::from_slice(&s[..]).map_err(Into::into)
@@ -1548,6 +1601,7 @@ fn idl(cfg_override: &ConfigOverride, subcmd: IdlCommand) -> Result<()> {
             program_id,
             filepath,
         } => idl_init(cfg_override, program_id, filepath),
+        IdlCommand::Close { program_id } => idl_close(cfg_override, program_id),
         IdlCommand::WriteBuffer {
             program_id,
             filepath,
@@ -1586,6 +1640,17 @@ fn idl_init(cfg_override: &ConfigOverride, program_id: Pubkey, idl_filepath: Str
         let idl_address = create_idl_account(cfg, &keypair, &program_id, &idl)?;
 
         println!("Idl account created: {:?}", idl_address);
+        Ok(())
+    })
+}
+
+fn idl_close(cfg_override: &ConfigOverride, program_id: Pubkey) -> Result<()> {
+    with_workspace(cfg_override, |cfg| {
+        let idl_address = IdlAccount::address(&program_id);
+        idl_close_account(cfg, &program_id, idl_address)?;
+
+        println!("Idl account closed: {:?}", idl_address);
+
         Ok(())
     })
 }
@@ -1759,6 +1824,44 @@ fn idl_erase_authority(cfg_override: &ConfigOverride, program_id: Pubkey) -> Res
     }
 
     idl_set_authority(cfg_override, program_id, None, ERASED_AUTHORITY)?;
+
+    Ok(())
+}
+
+fn idl_close_account(cfg: &Config, program_id: &Pubkey, idl_address: Pubkey) -> Result<()> {
+    let keypair = solana_sdk::signature::read_keypair_file(&cfg.provider.wallet.to_string())
+        .map_err(|_| anyhow!("Unable to read keypair file"))?;
+    let url = cluster_url(cfg, &cfg.test_validator);
+    let client = RpcClient::new(url);
+
+    // Instruction accounts.
+    let accounts = vec![
+        AccountMeta::new(idl_address, false),
+        AccountMeta::new_readonly(keypair.pubkey(), true),
+        AccountMeta::new(keypair.pubkey(), true),
+    ];
+    // Instruction.
+    let ix = Instruction {
+        program_id: *program_id,
+        accounts,
+        data: { serialize_idl_ix(anchor_lang::idl::IdlInstruction::Close {})? },
+    };
+    // Send transaction.
+    let latest_hash = client.get_latest_blockhash()?;
+    let tx = Transaction::new_signed_with_payer(
+        &[ix],
+        Some(&keypair.pubkey()),
+        &[&keypair],
+        latest_hash,
+    );
+    client.send_and_confirm_transaction_with_spinner_and_config(
+        &tx,
+        CommitmentConfig::confirmed(),
+        RpcSendTransactionConfig {
+            skip_preflight: true,
+            ..RpcSendTransactionConfig::default()
+        },
+    )?;
 
     Ok(())
 }
@@ -2116,6 +2219,7 @@ fn test(
     detach: bool,
     tests_to_run: Vec<String>,
     extra_args: Vec<String>,
+    env_vars: Vec<String>,
     cargo_args: Vec<String>,
 ) -> Result<()> {
     let test_paths = tests_to_run
@@ -2142,6 +2246,7 @@ fn test(
                 BootstrapMode::None,
                 None,
                 None,
+                env_vars,
                 cargo_args,
                 false,
             )?;
@@ -2784,9 +2889,22 @@ fn create_idl_account(
 
     // Run `Create instruction.
     {
-        let data = serialize_idl_ix(anchor_lang::idl::IdlInstruction::Create {
-            data_len: (idl_data.len() as u64) * 2, // Double for future growth.
-        })?;
+        let pda_max_growth = 60_000;
+        let idl_header_size = 44;
+        let idl_data_len = idl_data.len() as u64;
+        // We're only going to support up to 6 instructions in one transaction
+        // because will anyone really have a >60kb IDL?
+        if idl_data_len > pda_max_growth {
+            return Err(anyhow!(
+                "Your IDL is over 60kb and this isn't supported right now"
+            ));
+        }
+        // Double for future growth.
+        let data_len = (idl_data_len * 2).min(pda_max_growth - idl_header_size);
+
+        let num_additional_instructions = data_len / 10000;
+        let mut instructions = Vec::new();
+        let data = serialize_idl_ix(anchor_lang::idl::IdlInstruction::Create { data_len })?;
         let program_signer = Pubkey::find_program_address(&[], program_id).0;
         let accounts = vec![
             AccountMeta::new_readonly(keypair.pubkey(), true),
@@ -2796,14 +2914,27 @@ fn create_idl_account(
             AccountMeta::new_readonly(*program_id, false),
             AccountMeta::new_readonly(solana_program::sysvar::rent::ID, false),
         ];
-        let ix = Instruction {
+        instructions.push(Instruction {
             program_id: *program_id,
             accounts,
             data,
-        };
+        });
+
+        for _ in 0..num_additional_instructions {
+            let data = serialize_idl_ix(anchor_lang::idl::IdlInstruction::Resize { data_len })?;
+            instructions.push(Instruction {
+                program_id: *program_id,
+                accounts: vec![
+                    AccountMeta::new(idl_address, false),
+                    AccountMeta::new_readonly(keypair.pubkey(), true),
+                    AccountMeta::new_readonly(solana_program::system_program::ID, false),
+                ],
+                data,
+            });
+        }
         let latest_hash = client.get_latest_blockhash()?;
         let tx = Transaction::new_signed_with_payer(
-            &[ix],
+            &instructions,
             Some(&keypair.pubkey()),
             &[&keypair],
             latest_hash,
@@ -3120,6 +3251,7 @@ fn login(_cfg_override: &ConfigOverride, token: String) -> Result<()> {
 fn publish(
     cfg_override: &ConfigOverride,
     program_name: String,
+    env_vars: Vec<String>,
     cargo_args: Vec<String>,
     skip_build: bool,
 ) -> Result<()> {
@@ -3250,6 +3382,7 @@ fn publish(
             BootstrapMode::None,
             None,
             None,
+            env_vars,
             cargo_args,
             true,
         )?;
@@ -3332,6 +3465,7 @@ fn localnet(
     skip_build: bool,
     skip_deploy: bool,
     skip_lint: bool,
+    env_vars: Vec<String>,
     cargo_args: Vec<String>,
 ) -> Result<()> {
     with_workspace(cfg_override, |cfg| {
@@ -3349,6 +3483,7 @@ fn localnet(
                 BootstrapMode::None,
                 None,
                 None,
+                env_vars,
                 cargo_args,
                 false,
             )?;